refactor(DB, Helpers, misc): replace C-style casts across remaining modules
- DB/query: Replace C-style cast with static_cast for string conversion
- DB_PostgreSQL/query_pgsql: Replace C-style casts with static_cast/reinterpret_cast
  for PostgreSQL type conversions and struct field access
- Helpers/strconv: Replace C-style cast with static_cast
- DataFormat_JWT/jwt.h: Replace C-style cast with static_cast
- Program_Service/application: Replace C-style cast with static_cast
- App_FastRPC1Client/rpcclientapplication: Replace C-style cast with static_cast
- Scripts_JSONExprEval/atomicexpressionside: Replace C-style cast with static_cast

+++ Mantids30/Program_Service/application.cpp
@@ -22,7 +22,6 @@
 #include <sys/time.h>
 #include <sys/types.h>
 
-#include <cstring>
 #include <unistd.h>
 
 #include <Mantids30/Helpers/mem.h>
@@ -325,7 +324,7 @@ int StartApplication(int argc, char *argv[], Application *_app)
         if (!globalArgs.isInifiniteWaitAtEnd())
         {
             // Finish up.
-            syslog(LOG_NOTICE, "terminated (%" PRId32 ") by program execution", (int32_t) r);
+            syslog(LOG_NOTICE, "terminated (%" PRId32 ") by program execution", r);
             closelog();
             return r;
         }
